requirements.txt
python  3.8.3
torch  1.6.0
CUDA  11.0
numpy~=1.18.5
pillow~=7.2.0
opencv-python~=4.4.0.46
matplotlib~=3.2.2
torchvision~=0.7.0
scikit-learn~=0.23.1
pandas~=1.0.5
imgaug~=0.4.0
scipy~=1.5.0
scikit-image~=0.16.2
tensorflow~=2.3.1
dlib~=19.21.1
